feat: Add raw data endpoint for assessments
This commit adds a new endpoint `/assessments/:assessment/data/:variable` to retrieve raw data for a specific assessment and variable. The code changes include adding a new method `allForAssessmentRaw` to the `AssessmentModel` that fetches the raw data from the database and returns it in the response. This endpoint will be used to fetch raw data for assessments in the frontend.

Signed-off-by: Dheshan Mohandass <dheshan@mohandass.com>